Fix BatchUpdater SQL query type mismatch causing 0 record updates
Remove unnecessary ::integer cast in UPDATE query WHERE clause. The cast was
causing a type mismatch that prevented flash_id matching, resulting in
"Updated 0 records successfully" even when records should be updated.

This fixes the consumer receiving RabbitMQ events but failing to update
database records with ipfs_cid values after successful IPFS uploads.

+++ src/util/batch-updater/index.ts
@@ -73,7 +73,7 @@ export class BatchUpdater {
           UPDATE flashes 
           SET ipfs_cid = updates.ipfs_cid
           FROM (VALUES ${values}) AS updates(flash_id, ipfs_cid)
-          WHERE flashes.flash_id = updates.flash_id::integer
+          WHERE flashes.flash_id = updates.flash_id
           AND flashes.ipfs_cid IS NULL
         `;
